+++ ECG/jp2Encoder.h
@@ -6,7 +6,7 @@ private:
 	jas_stream_t* pStream;  //the pointer to the allocated stream, which stores the encoded data
 	jas_image_t* pImage; //the image data structure which stores the raw ECG image
 	jas_matrix_t* ref_data;
-	int jp2Size;  //the size of the encoded data stored in the stream
+	unsigned long jp2Size;  //the size of the encoded data stored in the stream
 	unsigned char* p_jp2Data;
 
 	int width, height;
@@ -16,6 +16,6 @@ public:
 	
 	int Encode();
 	int writeCompon(int*, int, int);
-	int getjp2Size();
+	unsigned long getjp2Size();
 	unsigned char* getjp2Data();
 };
